refactor(cockpit): move proof evidence tests to owner modules
